Create objectMeta before creating service
* new file objectmetacreator with func to create ObjectMeta
* new file servicecreator with functions to create/update service.
* moved service code away from resourcecreator to servicecreator.
